@Xyene, since you insist...
"Professor Quirrell," Harry said, omitting the phrase your approach to dealing with obstacles has certain drawbacks, "what if somebody looks in the door and sees that the Cerberus is dead?"

"Then they have probably already noticed something wrong with Snape," said Professor Quirrell. "But since you insist..." The Defense Professor walked over to the three-headed corpse and placed his wand against it. He began a Latin-sounding incantation that was accompanied by a sense of rising apprehension, the Boy-Who-Lived feeling the Dark Lord's power as he always had.

The last word spoken was "Inferius" and it was accompanied by a final surge of STOP, DON'T.

And the three-headed dog rose to a stand, its six eyes dull and blank, turning to watch the door once more.